feat(frontend): initialize project with Vite, React, and Chakra UI

+++ backend/controllers/product.controller.js
@@ -74,6 +74,14 @@ export const updateProduct = async (req, res) => {
 //** DELETE PRODUCT */
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({
+      success: false,
+      message: "Invalid product id!",
+    });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({
@@ -81,9 +89,9 @@ export const deleteProduct = async (req, res) => {
       message: "Product deleted!",
     });
   } catch (error) {
-    res.status(404).json({
+    res.status(500).json({
       success: false,
-      message: "Product not found!",
+      message: "Server error!",
     });
   }
 };
